Bugfix and Improvements
Fixed Issues:

Fixed the index.php when "iPad" resolution using a better CSS and grid configuration.

Fixed an issue in which the mobile nav menu would move randomly when opening and closing the menu, the issued was resolved moving all the style to dynamically generate on JS and improving checks.

Fixed an issue which maximized the menu icon after increasing the resolution with opening the menu at least one time, the issued was resolved moving all the style to dynamically generate on JS and improving checks.

Improvements:

given style to the admin.php file using those of contact page.

Known Issues:
the footer still moves around when using non-standard resolution changes on desktop dragging around.